+++ static/html/index.html
@@ -106,7 +106,7 @@
       </div>
       <div class="colophon">
         <a href="https://github.com/5calls/5calls">
-          <i class="fa fa-github" aria-hidden="true"></i> openSource
+          <i class="fa fa-github" aria-hidden="true"></i> Open Source
         </a>
         <a href="https://twitter.com/make5calls">
           <i class="fa fa-twitter" aria-hidden="true"></i> Updates on Twitter
