+++ static/js/main.js
@@ -84,12 +84,10 @@ store.getAll('org.5calls.userlocale', (userLocale) => {
     logger.debug("user locale get", userLocale[0]);
     cachedUserLocale = userLocale[0];
   } else {
-    const cachedUserLocale = userLocaleDetection(navigator.language || navigator.userLanguage);
+    cachedUserLocale = userLocaleDetection(navigator.language || navigator.userLanguage);
     store.add('org.5calls.userlocale', cachedUserLocale, () => {});
   }
 });
-
-
 
 // get stored user stats
 const defaultStats = {
@@ -440,7 +438,11 @@ app.model({
 
       send('setUserStats', data, done);
 
-      const body = queryString.stringify({ location: state.zip, result: data.result, contactid: data.contactid, issueid: data.issueid })
+      // This parameter will indicate to the backend api where this call report came from
+      // A value of test indicates that it did not come from the production environment
+      const viaParameter = window.location.host === '5calls.org' ? 'web' : 'test';
+      
+      const body = queryString.stringify({ location: state.zip, result: data.result, contactid: data.contactid, issueid: data.issueid, via: viaParameter })
       http.post(appURL+'/report', { body: body, headers: {"Content-Type": "application/x-www-form-urlencoded"} }, () => {
         // don’t really care about the result
       })
