+++ static/js/utils/localization_test.js
@@ -1,4 +1,4 @@
-const userLocaleDetection = require('./userLocaleDetection.js');
+const l = require('./localization');
 const chai = require('chai');
 const expect = chai.expect;
 
@@ -7,32 +7,29 @@ describe('userLocaleDetection', () => {
     it('should get "en" locale when an undefined argument is passed', () => {
       let lang = undefined;
       let expected = 'en';
-      let result = userLocaleDetection(lang);
+      let result = l.getLocaleFromBrowserLanguage(lang);
       expect(result).to.equal(expected);
     });
 
     it('should get "en" locale when an unsupported language is passed', () => {
       let lang = 'na';
       let expected = 'en';
-      let result = userLocaleDetection(lang);
+      let result = l.getLocaleFromBrowserLanguage(lang);
       expect(result).to.equal(expected);
     });
 
     it('should get the supported locale when a country within a supported locale is passed', () => {
       let lang = 'es-mx';
       let expected = 'es';
-      let result = userLocaleDetection(lang);
+      let result = l.getLocaleFromBrowserLanguage(lang);
       expect(result).to.equal(expected);
     });
 
     it('should get the supported locale when the supported locale is passed', () => {
       let lang = 'es';
       let expected = 'es';
-      let result = userLocaleDetection(lang);
+      let result = l.getLocaleFromBrowserLanguage(lang);
       expect(result).to.equal(expected);
     });
   });
 });
-
-
-
